# apps/api/src/users/users.controller.ts
import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Param, Patch, Post, Put, Query, UseGuards } from '@nestjs/common';
import { UsersService } from './users.service';
import { QueryFilterDto } from './dto/query-filter.dto';
import { UpdateUserDto } from './dto/update-user.dto';
import { Roles } from 'src/auth/roles.decorator';
import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
import { RolesGuard } from 'src/auth/roles.guard';
import { ApiBearerAuth, ApiBody, ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';

@ApiTags('users')
@Controller('users')
export class UsersController {
    constructor(private readonly usersService: UsersService) { }


    @ApiOperation({ summary: 'Rota exclusiva para administradores' })
    @ApiResponse({ status: 200, description: 'Acesso concedido ao admin' })
    @ApiResponse({ status: 401, description: 'Não autorizado' })
    @ApiResponse({ status: 403, description: 'Acesso negado' })
    @ApiBearerAuth()
    @Get('admin')
    @HttpCode(HttpStatus.OK)
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    getAdminData() {
        return { message: 'Bem-vindo, Admin!' };
    }

    @ApiOperation({ summary: 'Listar todos os usuários com filtro e paginação' })
    @ApiQuery({ name: 'filter', required: false, description: 'Texto para filtrar usuários' })
    @ApiQuery({ name: 'page', required: false, description: 'Número da página para paginação' })
    @ApiResponse({ status: 200, description: 'Lista de usuários retornada com sucesso' })
    @ApiResponse({ status: 401, description: 'Não autorizado' })
    @ApiBearerAuth()
    @Get()
    @HttpCode(HttpStatus.OK)
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    findAll(@Query() query: QueryFilterDto) {
        return this.usersService.findAll(query.filter, query.page);
    }

    @ApiOperation({ summary: 'Buscar um usuário pelo Id' })
    @ApiResponse({ status: 200, description: 'Usuário encontrado com sucesso' })
    @ApiResponse({ status: 404, description: 'Usuário não encontrado' })
    @ApiBearerAuth()
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    @Get(':id')
    @HttpCode(HttpStatus.OK)
    findOne(@Param('id') id: string) {
        return this.usersService.findOne(id);
    }

    @ApiBody({ type: UpdateUserDto })
    @ApiOperation({ summary: 'Atualizar completamente um usuário pelo ID' })
    @ApiResponse({ status: 200, description: 'Usuário atualizado com sucesso' })
    @ApiResponse({ status: 400, description: 'Dados inválidos' })
    @ApiResponse({ status: 404, description: 'Usuário não encontrado' })
    @ApiBearerAuth()
    @Put(':id')
    @HttpCode(HttpStatus.OK)
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    update(@Param('id') id: string, @Body() body: UpdateUserDto) {
        return this.usersService.update(id, body);
    }

    @ApiBody({ type: UpdateUserDto })
    @ApiOperation({ summary: 'Atualizar parcialmente um usuário pelo Id' })
    @ApiResponse({ status: 200, description: 'Usuário atualizado parcialmente com sucesso' })
    @ApiResponse({ status: 400, description: 'Dados inválidos' })
    @ApiResponse({ status: 404, description: 'Usuário não encontrado' })
    @ApiBearerAuth()
    @Patch(':id')
    @HttpCode(HttpStatus.OK)
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    partialUpdate(@Param('id') id: string, @Body() body: Partial<UpdateUserDto>) {
        return this.usersService.update(id, body);
    }

    @ApiOperation({ summary: 'Remover um usuário pelo Id' })
    @ApiResponse({ status: 204, description: 'Usuário removido com sucesso' })
    @ApiResponse({ status: 404, description: 'Usuário não encontrado' })
    @ApiBearerAuth()
    @UseGuards(JwtAuthGuard, RolesGuard)
    @Roles('ADMIN')
    @Delete(':id')
    @HttpCode(HttpStatus.NO_CONTENT)
    remove(@Param('id') id: string): void {
        this.usersService.remove(id);
    }
}