# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/admin": {
      "get": {
        "operationId": "UsersController_getAdminData",
        "summary": "Rota exclusiva para administradores",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Acesso concedido ao admin"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Listar todos os usuários com filtro e paginação",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Texto para filtrar usuários",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página para paginação",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Buscar um usuário pelo Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "summary": "Atualizar completamente um usuário pelo ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_partialUpdate",
        "summary": "Atualizar parcialmente um usuário pelo Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado parcialmente com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Remover um usuário pelo Id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário removido com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Lennon",
            "description": "Nome do usuário"
          },
          "email": {
            "type": "string",
            "example": "example@email.com",
            "description": "Email do usuário"
          },
          "password": {
            "type": "string",
            "example": "123456",
            "description": "Senha do usuário"
          }
        }
      },
...
    }
  }
}